extract TaxonomicBreakdownPopover
