+++ frontend/src/scenes/insights/filters/BreakdownFilter/TaxonomicBreakdownButton.tsx
@@ -1,50 +1,18 @@
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
 import { useState } from 'react'
-import { Popover } from 'lib/lemon-ui/Popover/Popover'
-import { TaxonomicFilter } from 'lib/components/TaxonomicFilter/TaxonomicFilter'
-import { useActions, useValues } from 'kea'
-import { groupsModel } from '~/models/groupsModel'
-import { insightLogic } from 'scenes/insights/insightLogic'
+import { useValues } from 'kea'
 import { LemonButton } from '@posthog/lemon-ui'
 import { IconPlusMini } from 'lib/lemon-ui/icons'
 import { taxonomicBreakdownFilterLogic } from './taxonomicBreakdownFilterLogic'
+import { TaxonomicBreakdownPopover } from './TaxonomicBreakdownPopover'
 
 export function TaxonomicBreakdownButton(): JSX.Element {
     const [open, setOpen] = useState(false)
-    const { allEventNames } = useValues(insightLogic)
-    const { groupsTaxonomicTypes } = useValues(groupsModel)
 
-    const { taxonomicBreakdownType, includeSessions } = useValues(taxonomicBreakdownFilterLogic)
-    const { addBreakdown } = useActions(taxonomicBreakdownFilterLogic)
-
-    const taxonomicGroupTypes = [
-        TaxonomicFilterGroupType.EventProperties,
-        TaxonomicFilterGroupType.PersonProperties,
-        TaxonomicFilterGroupType.EventFeatureFlags,
-        ...groupsTaxonomicTypes,
-        TaxonomicFilterGroupType.CohortsWithAllUsers,
-        ...(includeSessions ? [TaxonomicFilterGroupType.Sessions] : []),
-        TaxonomicFilterGroupType.HogQLExpression,
-    ]
+    const { taxonomicBreakdownType } = useValues(taxonomicBreakdownFilterLogic)
 
     return (
-        <Popover
-            overlay={
-                <TaxonomicFilter
-                    groupType={taxonomicBreakdownType}
-                    onChange={(taxonomicGroup, value) => {
-                        if (value) {
-                            addBreakdown(value, taxonomicGroup)
-                            setOpen(false)
-                        }
-                    }}
-                    eventNames={allEventNames}
-                    taxonomicGroupTypes={taxonomicGroupTypes}
-                />
-            }
-            visible={open}
-            onClickOutside={() => setOpen(false)}
-        >
+        <TaxonomicBreakdownPopover open={open} setOpen={setOpen}>
             <LemonButton
                 type={'secondary'}
                 icon={<IconPlusMini color="var(--primary)" />}
@@ -57,6 +25,6 @@ export function TaxonomicBreakdownButton(): JSX.Element {
                     ? 'Add cohort'
                     : 'Add breakdown'}
             </LemonButton>
-        </Popover>
+        </TaxonomicBreakdownPopover>
     )
 }
