+++ frontend/src/scenes/insights/filters/BreakdownFilter/TaxonomicBreakdownPopover.tsx
@@ -1,0 +1,55 @@
+import { useActions, useValues } from 'kea'
+
+import { groupsModel } from '~/models/groupsModel'
+import { insightLogic } from 'scenes/insights/insightLogic'
+import { taxonomicBreakdownFilterLogic } from './taxonomicBreakdownFilterLogic'
+
+import { Popover } from 'lib/lemon-ui/Popover/Popover'
+import { TaxonomicFilter } from 'lib/components/TaxonomicFilter/TaxonomicFilter'
+import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
+
+type TaxonomicBreakdownPopoverProps = {
+    open: boolean
+    setOpen: (open: boolean) => void
+    children: React.ReactElement
+}
+
+export const TaxonomicBreakdownPopover = ({ open, setOpen, children }: TaxonomicBreakdownPopoverProps): JSX.Element => {
+    const { allEventNames } = useValues(insightLogic)
+    const { groupsTaxonomicTypes } = useValues(groupsModel)
+    const { taxonomicBreakdownType, includeSessions } = useValues(taxonomicBreakdownFilterLogic)
+
+    const { addBreakdown } = useActions(taxonomicBreakdownFilterLogic)
+
+    const taxonomicGroupTypes = [
+        TaxonomicFilterGroupType.EventProperties,
+        TaxonomicFilterGroupType.PersonProperties,
+        TaxonomicFilterGroupType.EventFeatureFlags,
+        ...groupsTaxonomicTypes,
+        TaxonomicFilterGroupType.CohortsWithAllUsers,
+        ...(includeSessions ? [TaxonomicFilterGroupType.Sessions] : []),
+        TaxonomicFilterGroupType.HogQLExpression,
+    ]
+
+    return (
+        <Popover
+            overlay={
+                <TaxonomicFilter
+                    groupType={taxonomicBreakdownType}
+                    onChange={(taxonomicGroup, value) => {
+                        if (value) {
+                            addBreakdown(value, taxonomicGroup)
+                            setOpen(false)
+                        }
+                    }}
+                    eventNames={allEventNames}
+                    taxonomicGroupTypes={taxonomicGroupTypes}
+                />
+            }
+            visible={open}
+            onClickOutside={() => setOpen(false)}
+        >
+            {children}
+        </Popover>
+    )
+}
